add updates to support pkb_scheduler features
fix netperf by reverting to master

add flag for netperf rr test length

fix command to use netperf_rr_test_length when appropriate

add import vm_util to ping

reupdate to f string

fix bug

bug in ntttcp benchmark, removed it from metabenchmark until it can be fixed

add sleep to throughput_latency_jitter

change netperf start port

add code to update name

update throughput_latency_jitter

add updates to support pkb_scheduler features

fix command to use netperf_rr_test_length when appropriate

reupdate to f string

fix bug

bug in ntttcp benchmark, removed it from metabenchmark until it can be fixed

add sleep_time and interval flags to iperf

make regex more robust

add mtu flag for GCP

add checking mtu to iperf benchmark

initial commit for netperf multi-endpoint test

make test work

update test config

add vpc_id and subnet_id to config for gcp machines

small performance updates

add benchmark that does nothing

add wmem/rmem default flag

add inbound test

add netperf aggregate stuff

change to beta

perfkitbenchmarker/pkb.py

fix bug

fix bug

fix another bug

increase timeout

add netstat stats to netperf benchmark

add to metadata

updates to netperf and netperf aggregate benchmarks

align with master

remove netstat stuff

increase timeout

add date_time to pkb

add aws ga support

update metadata in publisher

update netperf to change sending and receiving to perform in both directions sequentially

add local bytes sent to tcp stream in netperf

nping update

add tcp retransmission metrics to netperf multistream

remove metric packets. Isn't accurate for TCP_RR, need to take send_size into account

add retry to iperf interval stuff

add iperf autostop benchmark

double to float

change enum to list for omb benchmark

fix imb bug for missing patch

Revert "fix imb bug for missing patch"

This reverts commit 584011fe9d38c765271e987d77c373a5aa7fa973.

add flags to increase tcp memory

reduce time between ping runs

add skip sysctl flag

revert to not all permutations for netperf

skip cleanup if skip prepare

skip cleanup/teardown for static vms

+++ perfkitbenchmarker/windows_virtual_machine.py
@@ -42,6 +42,10 @@ flags.DEFINE_bool(
     'set_cpu_priority_high', False,
     'Allows executables to be set to High (up from Normal) CPU priority '
     'through the SetProcessPriorityToHigh function.')
+flags.DEFINE_bool(
+    'skip_package_cleanup', False,
+    'skips cleanup')
+
 
 # Windows disk letter starts from C, use a larger disk letter for attached disk
 # to avoid conflict. On Azure, D is reserved for DvD drive.
@@ -456,7 +460,7 @@ class BaseWindowsMixin(virtual_machine.BaseOsMixin):
     self.home_dir = stdout.strip()
     stdout, _ = self.RemoteCommand('echo $env:SystemDrive')
     self.system_drive = stdout.strip()
-    self.RemoteCommand('mkdir %s' % self.temp_dir)
+    self.RemoteCommand('mkdir %s' % self.temp_dir, ignore_failure=True)
     self.DisableGuestFirewall()
 
   def _Reboot(self):
@@ -502,8 +506,11 @@ class BaseWindowsMixin(virtual_machine.BaseOsMixin):
     Deletes the Perfkit Benchmarker temp directory on the VM
     and uninstalls all PerfKit packages.
     """
+    if FLAGS.skip_package_cleanup:
+      return
     for package_name in self._installed_packages:
       self.Uninstall(package_name)
+
     self.RemoteCommand('rm -recurse -force %s' % self.temp_dir)
     self.EnableGuestFirewall()
